Moved all cumulus indexing implementation files to base project
