--- collections-online-cumulus/indexing/transformations/index.js
@@ -1,13 +1,0 @@
-module.exports = [
-  require('./field-names'),
-  require('./empty-title'),
-  require('./dates'),
-  require('./date-intervals'),
-  require('./relations'),
-  require('./dimensions'),
-  require('./geo-code'),
-  require('./latitude-longitude'),
-  require('./split-tags'),
-  require('./vision-tags'),
-  require('./tag-hierarchy')
-];
