+++ indexing/initialize/mode.js
