+++ indexing/modes/clear.js
