+++ indexing/post-processing/print-asset-exceptions.js
