+++ indexing/processing/result.js
@@ -5,9 +5,9 @@
  */
 
 const _ = require('lodash');
-const cip = require('../../../services/cip');
-const config = require('../../../collections-online/lib/config');
-const es = require('../../../collections-online/lib/services/elasticsearch');
+const cip = require('../../services/cip');
+const config = require('../../collections-online/lib/config');
+const es = require('../../collections-online/lib/services/elasticsearch');
 const Q = require('q');
 
 function AssetIndexingError(catalogAlias, assetId, innerError) {
