+++ indexing/transformations/dimensions.js
